requirements.txt
google-generativeai
langchain
langchain-core
langchain-chroma
langchain-community
chromadb
pypdf
python-dotenv
langchain_google_genai
pypdf2
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3